chore: changed some paths, updated funding table
